# api_gateway/docs/swagger.yaml
basePath: /
definitions:
...
  voting.CreatePublicVoteReq:
    properties:
      candidate_id:
        type: string
      election_id:
        type: string
      public_id:
        type: string
    type: object
...
  voting.GetAllVotes:
    properties:
      votes:
        items:
          $ref: '#/definitions/voting.GetVoteById'
        type: array
    type: object
...
  voting.GetPublicVoteRes:
    properties:
      candidate_id:
        type: string
      election_id:
        type: string
      id:
        type: string
      public_id:
        type: string
    type: object
  voting.GetVoteById:
    properties:
      candidate_id:
        type: string
      count:
        type: integer
    type: object
  voting.Void:
    type: object
info:
  contact: {}
  title: Online Voting System Swagger UI
paths:
...
  /public_vote/create:
    post:
      consumes:
      - application/json
      description: Endpoint for creating a new public_vote
      parameters:
      - description: Public_vote creation request payload
        in: body
        name: public_vote
        required: true
        schema:
          $ref: '#/definitions/voting.CreatePublicVoteReq'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created public_vote
          schema:
            $ref: '#/definitions/voting.Void'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Failed to create public_vote
          schema:
            type: string
      summary: Create a new public_vote
      tags:
      - Public_vote
  /public_vote/getall:
    get:
      consumes:
      - application/json
      description: Endpoint for getting public_votes
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully getted public_votes
          schema:
            $ref: '#/definitions/voting.GetPublicVoteRes'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Failed to get public_votes
          schema:
            type: string
      summary: Get Public_votes
      tags:
      - Public_vote
  /public_vote/getbyid:
    get:
      consumes:
      - application/json
      description: Endpoint for getting public_vote
      parameters:
      - in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully getted public_vote
          schema:
            $ref: '#/definitions/voting.GetPublicVoteRes'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Failed to get public_vote
          schema:
            type: string
      summary: Get Public_vote
      tags:
      - Public_vote
  /vote/getall:
    get:
      consumes:
      - application/json
      description: Endpoint for getting vote
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully getted vote
          schema:
            $ref: '#/definitions/voting.GetAllVotes'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Failed to get vote
          schema:
            type: string
      summary: Get Votes
      tags:
      - Vote
  /vote/getbyid:
    get:
      consumes:
      - application/json
      description: Endpoint for getting vote
      parameters:
      - in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully getted vote
          schema:
            $ref: '#/definitions/voting.GetVoteById'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Failed to get vote
          schema:
            type: string
      summary: Get Vote
      tags:
      - Vote
securityDefinitions:
  BearerAuth:
    in: header
    name: role
    type: apiKey
swagger: "2.0"